Feat: carousel 사이즈 수정
