+++ src/components/Category/index.tsx
@@ -1,27 +1,12 @@
 import Link from 'next/link';
-import { useEffect, useState } from 'react';
 
 import styles from '@/components/Category/index.module.scss';
 import { CATEGORY_MENU, CATEGORY_MENU_MOBILE } from '@/libs/constants';
+import useIsMobile from '@/libs/hooks/useIsMobile';
 import useManageCategory from '@/libs/hooks/useManageCategory';
 
 export default function Category() {
-  const [isMobile, setIsMobile] = useState(false);
-  useEffect(() => {
-    const handleResize = () => {
-      const aspectRatio = window.innerWidth / window.innerHeight;
-      if (aspectRatio <= 3 / 2) {
-        setIsMobile(true);
-      } else {
-        setIsMobile(false);
-      }
-    };
-
-    window.addEventListener('resize', handleResize);
-    handleResize();
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
+  const { isMobile } = useIsMobile();
   const { hideMenu, mouseEvent } = useManageCategory(isMobile);
 
   const sizeBaseCategoryMenu = isMobile ? CATEGORY_MENU_MOBILE : CATEGORY_MENU;
